init day3

+++ day1/src/main.rs
@@ -1,27 +1,27 @@
 const _INPUT: &str = include_str!("input");
 const _EXAMPLE: &str = include_str!("example");
 
-fn _part1() -> u64 {
+fn _part1() -> usize {
     _INPUT
         .split("\n\n")
         .map(|elf| {
             elf.lines()
-                .filter_map(|food| food.parse::<u64>().ok())
+                .filter_map(|food| food.parse::<usize>().ok())
                 .sum()
         })
         .max()
         .unwrap_or(0)
 }
 
-fn _part2() -> u64 {
+fn _part2() -> usize {
     let mut sums = _INPUT
         .split("\n\n")
         .map(|elf| {
             elf.lines()
-                .filter_map(|food| food.parse::<u64>().ok())
+                .filter_map(|food| food.parse::<usize>().ok())
                 .sum()
         })
-        .collect::<Vec<u64>>();
+        .collect::<Vec<usize>>();
     sums.sort_unstable();
     sums.iter().rev().take(3).sum()
 }
